Step 5-7
"Step 5. Examine the method moveDisk() in RevesActionThread. It  already has code that displays the move in the window and waits  for a step. Don’t change this code.
Step 6. Finish the moveDisk() method by adding code that will remove the top disk from one pole and then add it to the other pole.
Step 7. To test this, in the executeApplication() method of RevesActionThread, add the line of code."