Step 10 - 12
"Step 10. Refer to the algorithm for finding the value of k and create a method to compute it in RevesActionThread.
Step 11. Refer to the recursive design from the pre-lab exercises and create the method reves() in RevesActionThread. Make sure to call the method that was just completed in the previous step.
Step 12. Change the single line of code in the executeApplication() method of RevesActionThread to call reves(). The source pole will be a, the destination will be d, and the extras will be b and c. Use the variable disks for the number of disks to move."